+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 - Bugfix Z21 service big endian loco addresses ([#81](https://github.com/OpenRemise/Frontend/pull/81))
+- Bugfix Z21 service `LAN_RAILCOM_GETDATA` uses little endian loco address ([#95](https://github.com/OpenRemise/Frontend/pull/95))
 
 ## 0.4.0
 - Add connection status ([#91](https://github.com/OpenRemise/Frontend/pull/91))